Got the DB up and running
I still need to add dummy data to it, but all the tables should be
extant and have the bulk of the constraints in place.

I also added a Notes file with information about the database and VMArgs
I copied and pasted the contents of the database into a file called
RentMeDatabaseScript becauseI want us to have a quick reference to it if
we don't want to go back and forth to the workbench.