Successfully pulling data from DB
Had to alter the database a little to accomodate what we need due to the
inclusion of the Personal Information fields and me not correctly
formatting members, I think. The result set is being obtained, now I
have to populate the fields in the alter table and we should be almost
there.